requirements.txt
Django==1.9.7
dj-database-url==0.4.1
dj-static==0.0.6
django-cors-headers==1.1.0
django-rest-swagger==0.3.7
django-storages==1.4.1
djangorestframework==3.3.3
gunicorn==19.6.0
psycopg2==2.6.1
six==1.10.0
sqlparse==0.1.19
static==1.1.1
static3==0.7.0
django-jet
django-summernote
